fix some clippy warning; update halo2 to develop
